- change type of column amount in expenses to numeric - make /expenses endpoint return all expenses from db, ordered by date (desc)

+++ script.sql
@@ -5,7 +5,7 @@ name CHARACTER VARYING (100));
 CREATE TABLE IF NOT EXISTS expenses
 (id SERIAL PRIMARY KEY,
 comment CHARACTER VARYING (100),
-amount INTEGER,
+amount NUMERIC,
 category_id INTEGER,
 expense_timestamp TIMESTAMP,
 CONSTRAINT FK_EXPENSE_ON_CATEGORY_ID FOREIGN KEY (category_id) REFERENCES categories (id));
